+++ sickrage/core/queues/show.py
@@ -287,18 +287,18 @@ class QueueItemAdd(ShowQueueItem):
 
             # this usually only happens if they have an NFO in their show dir which gave us a Indexer ID that has no proper english version of the show
             if not getattr(s, 'seriesname'):
-                sickrage.app.log.error(
+                sickrage.app.log.warning(
                     "Show in {} has no name on {}, probably the wrong language used to search with".format(self.showDir,
                                                                                                            index_name))
                 sickrage.app.alerts.error(_("Unable to add show"),
-                                          _(
-                                              "Show in {} has no name on {}, probably the wrong language. Delete .nfo and add manually in the correct language").format(
-                                              self.showDir, index_name))
+                                          _("Show in {} has no name on {}, probably the wrong language. Delete .nfo "
+                                            "and add manually in the correct language").format(self.showDir,
+                                                                                               index_name))
                 return self._finishEarly()
 
             # if the show has no episodes/seasons
             if not len(s):
-                sickrage.app.log.error("Show " + str(s['seriesname']) + " is on " + str(
+                sickrage.app.log.warning("Show " + str(s['seriesname']) + " is on " + str(
                     IndexerApi(self.indexer).name) + " but contains no season/episode data.")
                 sickrage.app.alerts.error(_("Unable to add show"),
                                           _("Show ") + str(s['seriesname']) + _(" is on ") + str(
@@ -370,7 +370,7 @@ class QueueItemAdd(ShowQueueItem):
                     #     self.show.sports = 1
 
         except indexer_exception as e:
-            sickrage.app.log.error(
+            sickrage.app.log.warning(
                 "Unable to add show due to an error with " + IndexerApi(
                     self.indexer).name + ": {}".format(e))
             if self.show: