+++ sickrage/core/searchers/daily_searcher.py
@@ -141,7 +141,7 @@ class DailySearcher(object):
                 if curStatus not in (WANTED, DOWNLOADED, SNATCHED, SNATCHED_PROPER):
                     continue
 
-                if curStatus != WANTED and curQuality != Quality.UNKNOWN:
+                if curStatus != WANTED:
                     if bestQualities:
                         if curQuality in bestQualities or curQuality > max(bestQualities):
                             continue