+++ sickrage/notifiers/libnotify.py
@@ -92,19 +92,19 @@ class LibnotifyNotifier(Notifiers):
         self.gobject = GObject
         return True
 
-    def _notify_snatch(self, ep_name):
+    def notify_snatch(self, ep_name):
         if sickrage.app.config.libnotify_notify_onsnatch:
             self._notify(self.notifyStrings[self.NOTIFY_SNATCH], ep_name)
 
-    def _notify_download(self, ep_name):
+    def notify_download(self, ep_name):
         if sickrage.app.config.libnotify_notify_ondownload:
             self._notify(self.notifyStrings[self.NOTIFY_DOWNLOAD], ep_name)
 
-    def _notify_subtitle_download(self, ep_name, lang):
+    def notify_subtitle_download(self, ep_name, lang):
         if sickrage.app.config.libnotify_notify_onsubtitledownload:
             self._notify(self.notifyStrings[self.NOTIFY_SUBTITLE_DOWNLOAD], ep_name + ": " + lang)
 
-    def _notify_version_update(self, new_version="??"):
+    def notify_version_update(self, new_version="??"):
         if sickrage.app.config.use_libnotify:
             update_text = self.notifyStrings[self.NOTIFY_GIT_UPDATE_TEXT]
             title = self.notifyStrings[self.NOTIFY_GIT_UPDATE]