+++ sickrage/notifiers/nmj.py
@@ -88,19 +88,19 @@ class NMJNotifier(Notifiers):
 
         return True
 
-    def _notify_snatch(self, ep_name):
+    def notify_snatch(self, ep_name):
         return False
         # Not implemented: Start the scanner when snatched does not make any sense
 
-    def _notify_download(self, ep_name):
+    def notify_download(self, ep_name):
         if sickrage.app.config.use_nmj:
             self._notifyNMJ()
 
-    def _notify_subtitle_download(self, ep_name, lang):
+    def notify_subtitle_download(self, ep_name, lang):
         if sickrage.app.config.use_nmj:
             self._notifyNMJ()
 
-    def _notify_version_update(self, new_version):
+    def notify_version_update(self, new_version):
         return False
         # Not implemented, no reason to start scanner.